test with a full log file from 5.3.1
Use the Dockerfile to build and test, including pymapd.

Run python/pymapd script that connects to the test db and runs queries.
This consistent workflow across versions of omnisci
lets us test with changes to the logs.
The script also normalizes timestamps in the log file to minimize diff.
This script should not be run with every commit,
only when omnisci version changes or workflow steps are added.
Test with omnisci 5.3.1.

Change the test_load_db script from shell/omnisql to python/pymapd.

Change test_lines.rs to only test a single line, not load from a file.